Update seeders and migrations: consolidate schema changes, remove redundant migration files, and adjust footer view

+++ database/migrations/2025_06_21_000000_create_cart_items_table.php
@@ -4,14 +4,19 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
-    public function up(): void {
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
         Schema::create('cart_items', function (Blueprint $table) {
             $table->id();
             $table->string('session_id'); // Untuk guest users
             $table->foreignId('user_id')->nullable()->constrained('users')->onDelete('cascade');
             $table->foreignId('product_id')->constrained('products')->onDelete('cascade');
-            $table->integer('quantity');
+            $table->decimal('quantity', 8, 2);
             $table->timestamps();
             
             // Prevent duplicate items in cart
@@ -20,7 +25,11 @@ return new class extends Migration {
         });
     }
 
-    public function down(): void {
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
         Schema::dropIfExists('cart_items');
     }
 };
